backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
bcrypt==4.1.2
PyJWT==2.8.0
python-multipart==0.0.6